+++ services/web/app/storage/routes/shelf.py
@@ -17,24 +17,75 @@ from .misc import move_entity_to_storage
 
 
 from ..models import (
+    FixedColdStorage,
     FixedColdStorageShelf,
     CryovialBox,
 )
 from ...sample.models import Sample
 
-from ..forms import SampleToEntityForm, BoxToShelfForm
-from ..views import ShelfView
-from ...misc import chunks
+from ..forms import SampleToEntityForm, BoxToShelfForm, NewShelfForm
+from ..views import ShelfView, BasicShelfView
 from ..enums import EntityToStorageTpye
+from uuid import uuid4
 
-
-@storage.route("/shelves/view/LIMBSHF-<id>")
+@storage.route("/shelves/LIMBSHF-<id>")
 @login_required
 def view_shelf(id):
     shelf = ShelfView(id)
-    # Conversion to make it renderable in a nice way.
     return render_template("storage/shelf/view.html", shelf=shelf)
 
+
+@storage.route("/lts/LIMBLTS-<lts_id>/add_shelf", methods=["GET", "POST"])
+@login_required
+def add_shelf(lts_id):
+    lts = (
+        db.session.query(FixedColdStorage)
+        .filter(FixedColdStorage.id == lts_id)
+        .first_or_404()
+    )
+
+    form = NewShelfForm()
+
+    if form.validate_on_submit():
+        shelf = FixedColdStorageShelf(
+            name=form.name.data,
+            # Generate a UUID :)
+            uuid=uuid4(),
+            description=form.description.data,
+            storage_id=lts_id,
+            author_id=current_user.id,
+        )
+
+        db.session.add(shelf)
+        db.session.commit()
+
+        return redirect(url_for("storage.view_lts", lts_id=lts_id))
+
+    return render_template("/storage/shelf/new.html", form=form, lts=lts)
+
+
+@storage.route("/shelves/LIMBSHF-<id>/edit", methods=["GET", "POST"])
+@login_required
+def edit_shelf(id):
+    shelf = BasicShelfView(id)
+
+    form = NewShelfForm()
+
+    if form.validate_on_submit():
+        s = db.session.query(FixedColdStorageShelf).filter(FixedColdStorageShelf.id == id).first_or_404()
+
+        s.name = form.name.data
+        s.description = form.description.data
+        s.author_id = current_user.id
+        db.session.commit()
+        flash("Shelf successfully updated.")
+
+        return redirect(url_for("storage.view_shelf", id=id))
+
+    form.name.data = shelf["name"]
+    form.description.data = shelf["description"]
+
+    return render_template("storage/shelf/edit.html", shelf=shelf, form=form)
 
 @storage.route("/shelves/LIMBSHF-<shelf_id>/assign_box", methods=["GET", "POST"])
 @login_required
